requirements.txt
fastapi
fastapi-cli #! apparently useful
uvicorn
# DB
asyncpg #async psycopg
tortoise-orm # async orm
tomlkit
aerich #for db migrations
# SECURITY
python-jose[cryptography]
passlib[bcrypt]
# CONFIG
pydantic[email] #for validation and type enforcement
dotenv
Faker
httpx  #for async (probably)
# TESTING
pytest-asyncio #for testing async views (probably)
pytest